+++ xmake/plugins/config-update/xmake.lua
@@ -199,6 +199,7 @@ set_menu {
     { category = "Other Configuration" },
     { nil,                             "debugger",  "kv", "auto",     "Set debugger" },
     { nil,                             "ccache",    "kv", true,       "Enable or disable the c/c++ compiler cache." },
+    { nil,                             "debug",     "kv", false,      "Enable or disable to print debug information." },
     { nil,                             "ccachedir", "kv", nil,        "Set the ccache directory." },
     { nil, "trybuild", "kv", nil, "Enable try-build mode and set the third-party buildsystem tool.",
         "e.g.",